+++ lintrules-gradle/build.gradle
@@ -1,0 +1,20 @@
+apply plugin: 'java-library'
+apply plugin: 'kotlin'
+
+dependencies {
+    implementation project(':lintrules')
+
+    compileOnly "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+
+    compileOnly 'com.android.tools.lint:lint-api:26.5.1'
+    compileOnly 'com.android.tools.lint:lint-checks:26.5.1'
+
+    testImplementation 'com.android.tools.lint:lint:26.5.1'
+    testImplementation 'com.android.tools.lint:lint-tests:26.5.1'
+}
+
+jar {
+    manifest {
+        attributes('Lint-Registry-v2': 'com.chesire.lintrules.LintRegistry')
+    }
+}
