+++ lintrules-gradle/src/main/java/com/chesire/lintrules/gradle/LintRegistry.kt
@@ -1,10 +1,10 @@
-package com.chesire.lintrules
+package com.chesire.lintrules.gradle
 
 import com.android.tools.lint.client.api.IssueRegistry
 import com.android.tools.lint.detector.api.CURRENT_API
 import com.android.tools.lint.detector.api.Issue
-import com.chesire.lintrules.issues.DuplicateDependency
-import com.chesire.lintrules.issues.LexicographicDependencies
+import com.chesire.lintrules.gradle.issues.DuplicateDependency
+import com.chesire.lintrules.gradle.issues.LexicographicDependencies
 
 /**
  * Custom registry to add custom issues.
