--- lintrules/src/main/java/com/chesire/lintrules/issues/IssueHolder.kt
@@ -1,11 +1,0 @@
-package com.chesire.lintrules.issues
-
-import com.android.tools.lint.detector.api.Issue
-
-interface IssueHolder {
-    val issue: Issue
-    /**
-     * Message to display when reporting on this issue.
-     */
-    val message: String
-}
